feat: Changed indicators screen, dialog instead of huge list feat: indicador new ednpoint integration
