+++ src/translations/en-US.json
@@ -804,7 +804,8 @@
       "descriptiveWay": "In a descriptive way, what progress has your action made during the period?",
       "indicatorBaseline": "Indicator baseline (attach supporting information)",
       "sameContactQuestion": "Is the contact information responsible for the indicator data the same as the contact information registered in Section 1 - General Registration?",
-      "indicatorGoal": "Indicator Goal"
+      "indicatorGoal": "Indicator Goal",
+      "add": "Add new indicator"
     },
     "form5": {
       "title": "Monitoring and follow-up of climate actions through the SINAMECC",
